backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Semantic vs. Keyword Search Playground — FastAPI Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "weaviate-client>=4.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
